feat: Prevent category deletion when containing posts
* if the category contains posts, throw NonEmptyCategoryException as a warning.

+++ wild/src/main/java/study/wild/service/CategoryService.java
@@ -7,15 +7,20 @@ import org.springframework.transaction.annotation.Transactional;
 import study.wild.domain.Category;
 import study.wild.dto.CategoryDto;
 import study.wild.dto.PostDto;
+import study.wild.exception.NonEmptyCategoryException;
 import study.wild.repository.CategoryRepository;
+import study.wild.repository.PostRepository;
 
 import java.util.List;
 import java.util.stream.Collectors;
 
+// TODO: CategoryPostService 분리할 것
 @Service
 @Transactional(readOnly = true)
 @RequiredArgsConstructor
 public class CategoryService {
+
+    private final PostRepository postRepository;
 
     private final CategoryRepository categoryRepository;
 
@@ -51,7 +56,13 @@ public class CategoryService {
 
     @Transactional
     public void deleteCategory(Long categoryId) {
+        if (hasPostInCategory(categoryId)) {
+            throw new NonEmptyCategoryException();
+        }
         categoryRepository.deleteById(categoryId);
     }
 
+    public boolean hasPostInCategory(Long categoryId) {
+        return !postRepository.findPostByCategoryIdAndDeleted(categoryId, false).isEmpty();
+    }
 }
